#16 refactor: dummy 데이터들 dummy 폴더에서 별도 관리
